COSMIT refactor linear classifiers
Factored out the common code to LinearSVC, LogisticRegression, Perceptron
and SGDClassifier into a LinearClassifierMixin.
